+++ rs/libmoq/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.6](https://github.com/moq-dev/moq/compare/libmoq-v0.5.5...libmoq-v0.5.6) - 2026-08-07
+
+### Added
+
+- *(libmoq)* add a client config handle, and advanced settings to the OBS plugin ([#2650](https://github.com/moq-dev/moq/pull/2650))
+
+### Other
+
+- drop Intel macOS release targets ([#2715](https://github.com/moq-dev/moq/pull/2715))
+
 ## [0.5.5](https://github.com/moq-dev/moq/compare/libmoq-v0.5.4...libmoq-v0.5.5) - 2026-08-06
 
 ### Fixed