+++ rs/moq-audio/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.17](https://github.com/moq-dev/moq/compare/moq-audio-v0.0.16...moq-audio-v0.0.17) - 2026-08-07
+
+### Added
+
+- *(cli)* add moq play ([#2697](https://github.com/moq-dev/moq/pull/2697))
+
 ## [0.0.16](https://github.com/moq-dev/moq/compare/moq-audio-v0.0.15...moq-audio-v0.0.16) - 2026-08-06
 
 ### Added