+++ rs/moq-transcode/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.9](https://github.com/moq-dev/moq/compare/moq-transcode-v0.0.8...moq-transcode-v0.0.9) - 2026-08-07
+
+### Other
+
+- updated the following local packages: moq-video
+
 ## [0.0.8](https://github.com/moq-dev/moq/compare/moq-transcode-v0.0.7...moq-transcode-v0.0.8) - 2026-08-06
 
 ### Fixed